Improved how .bf and .br look like. Improved .slot result calculation performance (because of .slottest). Some string changes

+++ src/Nadeko.Econ/Gambling/Betflip/BetflipResult.cs
@@ -3,5 +3,5 @@
 public readonly struct BetflipResult
 {
     public decimal Won { get; init; }
-    public int Side { get; init; }
+    public byte Side { get; init; }
 }
