+++ src/Nadeko.Econ/Gambling/Slot/SlotResult.cs
@@ -3,7 +3,7 @@
 public readonly struct SlotResult
 {
     public decimal Multiplier { get; init; }
-    public int[] Rolls { get; init; }
+    public byte[] Rolls { get; init; }
     public decimal Won { get; init; }
     public SlotWinType WinType { get; init; }
 }
